Store the sheet data in Student objects
Tidy up the code that makes the call with the Google Sheet, so that it creates Student Objects and stores each of these in the studentList of the Assessment object.

NOTE: The code is still in the wrong place, this will need to be called after the classes and number of students has been identified.